Move load-external-script into helpers
