+++ packages/client/src/helpers/load-external-script.ts
